requirements.txt
flask
anthropic
pdfplumber
python-docx
Pillow
python-dotenv
